External receiver updates
- Move Intent parsing to separate file
- Support finding note by query
  (instead of selecting by ID or path)
- Give properties in note info
